pyproject.toml
[tool.poetry]
name = "ludivina"
version = "1.0.0"
description = "Ludivina - Thermal Printer Bot for Telegram"
authors = ["Guido <guido.garcia.bernardo@gmail.com>", "Gustavo <gustavogb@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^13.9"
thermalprinter = "^0.2.0"
Pillow = "^8.4.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pylint = "^2.12.2"
pip-licenses = "^3.5.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
